# RFC: Iterate over smaller set for setdiff[!](a,b)