@StefanKarpinski 2018-09-09
Is the 2x length ration threshold empirically determined or is there some intuition behind it? Either way it's probably worth noting in a comment.