@laborg 2018-09-10
I made some statistics with Ints as elements, starting with 1:1 ratio and it appeared that a good threshold for those is around 0.8, but this depends on a couple of things: 
- size of both sets
- size of the conjuction of both sets
- cost of hash-lookup (`ht_keyindex()`)
- cost of removal (`_delete()`)

With those variables its hard to get to a generic threshold and therefore the conservative estimation of 0.5 (~things should not regress performance wise).
But what I saw is that _substantial_ performance improvement is available for large size discrepancies anyway.

Maybe it would help to implement a faster `filter_in_one_pass` to get rid of the many lookups...