@fredrikekre 2018-09-11
@nanosoldier `runbenchmarks(ALL, vs=":release-1.0")`