@fredrikekre 2018-09-11
Derp
@nanosoldier `runbenchmarks(ALL, vs=":master")`