@laborg 2018-09-11
I want to find a better solution which also includes the other set functions (symdiff, untion, intersect) plagued by this asymmetric behaviour - should I in the mean time close this PR?