@StefanKarpinski 2018-09-11
It seems better to me to have a partial solution until we can have a more complete one. You can always revert this later if it's easier.