@nanosoldier 2018-09-11
[Your benchmark job](https://github.com/JuliaLang/julia/pull/29048#issuecomment-420214871) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/980ce0b5f935c7b6b7d221256053ef7ee3aaf972/313f777_vs_e85af88/report.md). cc @ararslan