@nanosoldier 2018-09-11
[Your benchmark job](https://github.com/JuliaLang/julia/pull/29048#issuecomment-420214742) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/b54549a7b78c8cebce6174863b36891bf7356939/313f777_vs_dc40b6f/report.md). cc @ararslan