@mbauman 2018-09-11
Benchmarks look to be their usual noisy selves.  We do have a few `setdiff` benchmarks, but the arguments used don't appear to have large enough size disparities to show up in the results.  I'll tag this as "needs benchmark" even though you provided good benchmarks in the first post just so we don't forget to migrate them over to BaseBenchmarks.jl.  Great first PR!